relation tube edit - отдельный компонент
